LayoutTests/fast/forms/color/input-color-onchange-event-expected.txt

Test if change event fires properly when color chooser changes. Bug 66848

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


value changed to #ff0000
PASS input.value is "#ff0000"
